feat: cast into double health and damage fix: getting gens from fight session

+++ src/main/java/com/hacman/gomunkulevolv/abilities/VampireAbility.java
@@ -16,8 +16,8 @@ public class VampireAbility extends Ability{
     }
 
     private void restoreHealthByVampiring(Creature creature) {
-        int curHealth = creature.getCurHealth();
-        int damage = creature.getDamage();
+        double curHealth = creature.getCurHealth();
+        double damage = creature.getDamage();
         double restoredHealth = curHealth + damage * vampireRate;
         creature.setCurHealth((int) restoredHealth);
     }
